refactor(print-format-builder): extract useColumnResize composable
